agent/pyproject.toml
[tool.poetry]
name = "gpt_agent"
version = "0.1"
description = ""
authors = []
readme = "README.md"
packages = [{ include = "gpt_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "~=0.103"
uvicorn = "~=0.23"
aiofiles = "~=23.2"
langchain = "~=0.0"
openai = "~=0.28"
httpx = "~=0.24"
python-dotenv = "~=1.0"


[tool.poetry.group.dev.dependencies]
pytest = "~=7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
